generalize SeleniumTestCase for reuse elsewhere
